static extension : OnnxTransform, DnnImageFeaturizer (OnnxTransform)
